+++ src/Model/Customer.cs
@@ -9,7 +9,7 @@ namespace HelpScoutNet.Model
     public class Address
     {
         [DefaultValue(0)]
-        public int Id { get; set; }
+        public long Id { get; set; }
         public List<string> Lines { get; set; }
         public string City { get; set; }
         public string State { get; set; }
@@ -28,7 +28,7 @@ namespace HelpScoutNet.Model
     public class SocialProfile
     {
         [DefaultValue(0)]
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Value { get; set; }
         public string Type { get; set; }
     }
@@ -36,7 +36,7 @@ namespace HelpScoutNet.Model
     public class Email
     {
         [DefaultValue(0)]
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Value { get; set; }
         public string Location { get; set; }
     }
@@ -44,7 +44,7 @@ namespace HelpScoutNet.Model
     public class Phone
     {
         [DefaultValue(0)]
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Value { get; set; }
         public string Location { get; set; }
     }
@@ -52,7 +52,7 @@ namespace HelpScoutNet.Model
     public class Chat
     {
         [DefaultValue(0)]
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Value { get; set; }                
         [JsonConverter(typeof(StringEnumConverter))]
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
@@ -76,14 +76,14 @@ namespace HelpScoutNet.Model
     public class Website
     {
         [DefaultValue(0)]
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Value { get; set; }
     }
 
     public class Customer
     {
         [DefaultValue(0)]
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string PhotoUrl { get; set; }
